+++ src/components/theme/Register/Register.jsx
@@ -4,7 +4,7 @@
  */
 
 import React, { Component } from 'react';
-import Helmet from 'react-helmet';
+import { Helmet } from '@plone/volto/helpers';
 import { connect } from 'react-redux';
 import { compose } from 'redux';
 import { defineMessages, injectIntl } from 'react-intl';