+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Internal
 
 - Pin Guillotina docker image @sneridagh
+- Forked `react-helmet` since it seems unmaintained. Now it's a Named import in helpers. @sneridagh
 - Update internal dependencies, fix "unmet peer dependencies" console logs by adding the peer dependencies to the local dependencies @sneridagh
 - Update some dependencies, including: react-router, eslint engine and plugins/config and others @sneridagh